+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -7,7 +7,7 @@ import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.handler.HandlerMapper;
-import org.apache.coyote.http11.request.Request;
+import org.apache.coyote.http11.message.request.Request;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
