+++ tomcat/src/main/java/org/apache/coyote/http11/message/ContentType.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.http11.message;
 
 import java.util.Arrays;
