+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/Cookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.message.request;
 
 import java.util.Arrays;
 import java.util.Collections;
