+++ tomcat/src/main/java/org/apache/coyote/http11/message/request/QueryParameter.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.message.request;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -14,7 +14,7 @@ public class QueryParameter {
     public QueryParameter(final String uri) {
         final int index = uri.indexOf("?");
         final String queryString = uri.substring(index + 1);
-        final String[] splitedQueryStrings = queryString.split("&"); // account=gugu
+        final String[] splitedQueryStrings = queryString.split("&");
         for (final String str : splitedQueryStrings) {
             final int strIndex = str.indexOf("=");
             parameters.put(str.substring(0, strIndex), str.substring(strIndex + 1));
